Move SelfAssignedDeclarationInspectionTests for QuickFixes into relevant files under RD.QuickFixes
