+++ RubberduckTests/QuickFixes/IgnoreOnceQuickFixTests.cs
@@ -892,5 +892,41 @@ End Sub";
             Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
         }
 
+
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void SelfAssignedDeclaration_IgnoreQuickFixWorks()
+        {
+            const string inputCode =
+@"Sub Foo()
+    Dim b As New Collection
+End Sub";
+
+            const string expectedCode =
+@"Sub Foo()
+'@Ignore SelfAssignedDeclaration
+    Dim b As New Collection
+End Sub";
+
+            var builder = new MockVbeBuilder();
+            var project = builder.ProjectBuilder("VBAProject", ProjectProtection.Unprotected)
+                .AddComponent("MyClass", ComponentType.ClassModule, inputCode)
+                .Build();
+            var component = project.Object.VBComponents[0];
+            var vbe = builder.AddProject(project).Build();
+
+            var parser = MockParser.Create(vbe.Object);
+
+            parser.Parse(new CancellationTokenSource());
+            if (parser.State.Status >= ParserState.Error) { Assert.Inconclusive("Parser Error"); }
+
+            var inspection = new SelfAssignedDeclarationInspection(parser.State);
+            var inspectionResults = inspection.GetInspectionResults();
+
+            new IgnoreOnceQuickFix(parser.State, new[] { inspection }).Fix(inspectionResults.First());
+            Assert.AreEqual(expectedCode, parser.State.GetRewriter(component).GetText());
+        }
+
+
     }
 }
